requirements.txt
# WealthLens OSS - Python Dependencies
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==44.0.0
authlib==1.4.1

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# AI Integration
anthropic==0.42.0

# Utilities
python-dotenv==1.0.1
scipy==1.15.0
numpy==1.26.4
nanoid==2.0.0
mftool==3.0

# File handling
aiofiles==24.1.0

# CORS
starlette==0.41.3

# Production server
gunicorn==23.0.0

# Cache
redis==5.2.1
